Add a concept of flights
A flight is a collection of airplanes moving together. This should
make it more logical when working on the game play and for the user as
airplanes typically operate in small groups (flights). Flights are to
be dynamically created, so it will be possible to create and break
such up, all the way down to individual airplanes, if desired.