Add concept of actor visual
This is the visual representation which is going to be a sprite and
is prepared to be a multiplexed sprite, even though I am not sure
whether it will be needed in the end. However, with multiple actors
and representing aerodromes with sprites, it is not unlikely that
multiplexing will be needed.