Allow for more than one list of visuals
Ownership of the map visuals is moved to the play state.

A dogfight will define its own visuals of the participating airplanes
stacked up, with the idea that there will be two opposing sides lined
up and this is a separate visual compared to the map.

While it may seem a bit overkill to allow for sprite multiplexing, it
comes for free and doing it this way should have the benefit of having
a uniform way of displaying a set of visual actors.