Remove list of flights from play state
There is no node to connect them with. Adding such node to flight is a
problem as there can only be one first element and this is the node
that is linked from the map. Flights can also be linked from
aerodromes or dogfights, but they reuse the same node.

The visual actor list uses a node that is not first, but this is
solved with a back pointer.

The idea of having flights in the play state was most likely for
quickly given orders, but orders are also for dogfights and dogfights
did not have a list.

It seems easiest to either pick up units from the map (or possibly
from the actor visual list), which allows traversal in some fairly
predictable way based on the location of the unit. Filtering can be
used to find the relevant actors for giving orders.

Long story, but the flights list is removed for the stated reasons
from the play state.