Fix type nodes ordering
The actor kind enumeration is now ordered based on stacking
priority. This will put flights on top of ground units, simply
speaking. It is slightly more complicated than that, but the idea is
that it gives an appropriate priority to handle rendering multiple
units at the same location.

My thinking is that stacking can be shown with a tiny but of offset to
each others so that one can visually see that there are more than one
unit there. The most important one should be at the top, which is why
flights have the highest stacking priority.

Also fix flights which had a different kind of node with ordering (not
kind). These (kind and sequence/order number) are now merged into a
single node type.