Improve sprite order and stagger actors
Insert actors in the display list not only based on the y location,
but also based on their priority order. The idea is that flights go
above ground units.

If there are multiple actors in the same location, use the order and
gradually stagger them to give a visual impression that there are
player units obscured. This does not make them visible, but it
visually shows the amount of units in the same location.

Introduce a reduced sprite info record as we only need a few fields
from it (coordinates are always calculated). Also provide two sets of
sprite info records to allow for blinking alternative display of an
actor. The actual timer to toggle remains to be implemented.