Make it build again
Add back sprite structure, it is small overhead and makes it easier to
have it that way.

Also put flights back to playstate. The reason they are there (and not
on the map) is that they can be placed between hexes. The map is only
for actors that are placed on a specific hex. Another reason to have
flights owned by the playstate is that it makes traversal easier
when giving orders.

Dogfights are now part of playstate. The reason is that I expect it
makes it easier to give orders by having a simple traversal.
The only downside is that it (later) makes for another loop to build
the display list. On the other hand, it is already needed for flights,
so just extend it with another list.