Remove actors in a safer way
When memory is to be freed there is a problem with the actor display
list and synchronization. Rather than freeing memory immediate, keep a
list of nodes to be freed and do that later when it is safe.

The idea is to have a simple synchronization with an interrupt that
builds the actor visual display list. We simply unlink the actor from
whatever list owns it, but not the display list. We store the actor in
a list to be freed. Then after we have synchronized with the interrupt
that rebuilds the actor display list, we get the control back and know
that the freed nodes no longer are part of the display list and memory
can be safely recycled at this point.