Simplify synchronization
The interrupt builds the actor visuals list and a simple atomic flag
is used for the main code to tell the interrupt when it should update
the list. The main code will then wait until it is done before
resuming operation.

This greatly simplifies the synchronization as only a simple flag is
needed.