Speed up traversing visual actors
The display will traverse the rectangle of visible sectors, this can
be 300 iterations and it should be fairly sparse. The idea is to keep
a counter for the longer q dimension, if there are no visual actors in
an entire "q lane" we can quickly skip past it.

Fix the visual actor being added to the map as it still used the old
hashmap, not the new sector_data array.

Also fix a problem with aerodromes that added the wrong node as the
visual actor.

Finally change the type for adding/removing visual actors, as it is
more correct and makes it somewhat more type safe.